requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-api-core==1.16.0
-e git://github.com/googleapis/google-api-python-client.git@ef6420a803618f968f3f47f2dba5cc983dc438af#egg=google_api_python_client
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pdfrw==0.4
protobuf==3.11.3
psutil==5.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
uritemplate==3.0.0
urllib3==1.25.7
Werkzeug==0.16.0
